mathモジュールをまとめる
